requirements.txt
# Python 3.10+ recomendado (Ubuntu 22.04+)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
